caption/requirements.txt
flask
torch
numpy
pillow
torchvision